+++ processes-shell/utils.h
@@ -1,0 +1,16 @@
+#ifndef WISH_UTILS_H
+#define WISH_UTILS_H
+
+#include <stdbool.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+#include "types.h"
+
+void print_token(Token *, bool);
+
+void warn();
+
+void error();
+
+#endif
